requirements.txt
# Solana相关
solana-py==0.30.2
anchorpy==0.14.0
solders>=0.18.0

# 网络和API
aiohttp==3.9.1
requests==2.31.0
websockets>=11.0.0

# GRPC相关
grpcio==1.59.3
grpcio-tools==1.59.3
protobuf==4.25.1
yellowstone-grpc>=0.4.0

# 数据库
aiosqlite==0.19.0
sqlalchemy==2.0.23

# 数据处理
pandas==2.1.3
numpy==1.24.3
python-dateutil==2.8.2

# 工具和辅助
python-dotenv==1.0.0
tqdm==4.66.1
loguru==0.7.2
click==8.1.7

# 测试
pytest==7.4.3
pytest-asyncio==0.21.1

# 日志和配置
pyyaml>=6.0

# Node.js服务依赖 (需要在parser-service目录下通过npm安装)
# npm install express @debridge-finance/solana-transaction-parser @solana/web3.js

# 系统要求
# Node.js >= 14.0.0
# Python >= 3.9.0 